% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_table_across_dbs.R
\name{compare_table_across_dbs}
\alias{compare_table_across_dbs}
\title{Compare tables between two databases}
\usage{
compare_table_across_dbs(fram_db_1, fram_db_2, table_name)
}
\arguments{
\item{fram_db_1}{fram database connection to first database (used as reference for ratio calculations)}

\item{fram_db_2}{fram database connection to second database}

\item{table_name}{Name of table to compare}
}
\value{
tibble of comparisons, including run_id, as appropriate stock_id, age, fishery_id, time_step. \verb{$db1} and \verb{$db2} give the values in the database for column \verb{$metric}; \verb{$diff_exact} has db1-db2, and \verb{$diff_ratio} has db2/db1.
}
\description{
Compare tables between two databases
}
\examples{
\dontrun{
library(framrsquared)
fram_a = connect_fram_db("C:/Repos/fram multirun testing/Formal testing/coho_preseason_notamm_a.mdb",
read_only = TRUE,
quiet = TRUE)
fram_c = connect_fram_db("C:/Repos/fram multirun testing/Formal testing/coho_preseason_notamm_c.mdb",
                        read_only = TRUE,
                       quiet = TRUE)
compare_table_across_dbs(fram_a, fram_c, "Mortality")
disconnect_fram_db(fram_a)
disconnect_fram_db(fram_b)
}
}
