% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framqaqc-package.R
\docType{package}
\name{framqaqc-package}
\alias{framqaqc}
\alias{framqaqc-package}
\title{framqaqc: What the Package Does (One Line, Title Case)}
\description{
QAQC tools for comparing FRAM databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRAMverse/framqaqc}
  \item \url{https://framverse.github.io/framqaqc/}
  \item Report bugs at \url{https://github.com/FRAMverse/framqaqc/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
