% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparisons.R
\name{plot_comparisons_exact}
\alias{plot_comparisons_exact}
\title{Plot the values of the two databases against each other}
\usage{
plot_comparisons_exact(tab, round_digits = 0)
}
\arguments{
\item{tab}{tibble produced from \code{compare_table_across_dbs()}.}

\item{round_digits}{Digits to round to before identifying entries as "near-duplicate" entries. Defaults to 0 (e.g., round to the nearest fish). Appropriate to reduce when comparing tables in which values are much smaller.}
}
\value{
ggplot object
}
\description{
Creates faceted plots based on the output of \code{compare_table_across_dbs()}, with the values from one database plotted against the corresponding values from the other. If two databases are exactly the same, all points should fall on the 1:1 line; dashed 1:1 line included as reference. To improve performance, avoids plotting duplicate and near-duplicate entries. "near-duplicate" is controlled by argument \code{round_digits}.
}
