% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparisons.R
\name{plot_comparisons_ratio}
\alias{plot_comparisons_ratio}
\title{Plot the ratios of database comparisons}
\usage{
plot_comparisons_ratio(tab, round_digits = 0)
}
\arguments{
\item{tab}{tibble produced from \code{compare_table_across_dbs()}.}

\item{round_digits}{digits to round db1 to in order to avoid plotting near-duplicate points.}
}
\value{
ggplot object
}
\description{
Creates faceted plots based on the output of \code{compare_table_across_dbs()}, with the ratio of db2/db1 plotted against the values of db1. If two databases are exactly the same, all points should have a ratio of exactly 1; dashed horizontal line at 1 is included as reference. To improve performance, avoids plotting duplicate and near-duplicate entries. "near-duplicate" values of db1 is controlled by argument \code{round_digits}; ratio values within 0.001 are treated as near-duplicate.
}
