% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tables.R
\name{summarize_exact}
\alias{summarize_exact}
\title{Summarize exact differences from \code{compare_table_across_dbs()}}
\usage{
summarize_exact(tab)
}
\arguments{
\item{tab}{tibble produced from \code{compare_table_across_dbs()}.}
}
\value{
gt with summary information.
}
\description{
Makes a quick summary of the outputs of compare_table_across_dbs(), giving
min, max, 0.01 and 0.99 quantiles of the absolute differences of values by run id. A perfect match would have values that are all 0.
}
