% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tables.R
\name{summarize_ratio}
\alias{summarize_ratio}
\title{Summarize ratio differences from \code{compare_table_across_dbs()}}
\usage{
summarize_ratio(tab)
}
\arguments{
\item{tab}{tibble produced from \code{compare_table_across_dbs()}.}
}
\value{
gt with summary information.
}
\description{
Makes a quick summary of the outputs of compare_table_across_dbs(), giving
min, max, 0.01 and 0.99 quantiles of the ratio of values from db2 / db1, by run id.
Ignore entries with values <=1 to db1 to avoid divide-by-zero issues and inflation of small differences.
}
